requirements.txt
alembic~=1.13.1
fastapi>=0.95.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.25.1
requests>=2.31.0
pydantic~=2.5.3
sqlalchemy~=2.0.28
pyodbc~=5.1.0
redis~=5.0.2
passlib~=1.7.4
python-jose~=3.3.0
numpy~=1.26.4
apscheduler~=3.10.4
pandas~=2.2.1
pymysql~=1.1.0
dbutils~=3.1.0
oss2~=2.18.4
loguru~=0.7.2
